Fix for safari desktop
On Safari for Macos the user menu is clipped due to a Safari bug. If you set both overflow:hidden and position:fixed on a parent element, a child element with position: fixed will incorrectly be clipped by the parent element. The fix for this is to remove overflow:visible on large screens, where it isn't required anyway.